requirements.txt
# Tekshila Backend - 2026 Ready Agent Architecture
# Python 3.11+ required

# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.17
python-dotenv>=1.0.1
pydantic>=2.9.0
pydantic-settings>=2.6.0

# Database
sqlalchemy>=2.0.36
alembic>=1.14.0
asyncpg>=0.30.0
redis>=5.2.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
authlib>=1.3.2
httpx>=0.27.0

# GitHub Integration
PyGithub>=2.5.0

# AI/ML Stack - Agent Architecture
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0
langsmith>=0.1.0

# Vector Store & Embeddings
chromadb>=0.5.0
sentence-transformers>=3.3.0
tiktoken>=0.8.0

# Code Analysis
tree-sitter>=0.23.0
tree-sitter-python>=0.23.0
tree-sitter-javascript>=0.23.0
tree-sitter-typescript>=0.23.0
radon>=6.0.1
pylint>=3.3.0

# Async & Concurrency
aiofiles>=24.1.0
aioredis>=2.0.1
celery>=5.4.0

# Observability
opentelemetry-api>=1.28.0
opentelemetry-sdk>=1.28.0
opentelemetry-instrumentation-fastapi>=0.49b0
structlog>=24.4.0

# Security
itsdangerous>=2.2.0
cryptography>=43.0.0

# Utilities
orjson>=3.10.0
ujson>=5.10.0
pyyaml>=6.0.2
tenacity>=9.0.0
httpx-sse>=0.4.0
